malloc checked need to clean up

+++ libft/includes/libft/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/13 13:16:59 by abidolet          #+#    #+#             */
-/*   Updated: 2025/05/31 19:10:10 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/04 23:34:21 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include <stdlib.h>
 # include <unistd.h>
-# include "str.h"
+# include "libft/str.h"
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
@@ -23,9 +23,10 @@
 
 # define FD_MAX 1024
 
-char	*get_next_line(int fd);
-ssize_t	ft_getdelim(char **restrict lineptr, size_t *restrict n, int delim,
-			int fd);
-ssize_t	ft_getline(char **restrict lineptr, size_t *restrict n, int fd);
+char			*get_next_line(int fd);
+unsigned long	ft_getdelim(char **restrict lineptr, unsigned long *restrict n,
+					int delim, int fd);
+unsigned long	ft_getline(char **restrict lineptr, unsigned long *restrict n,
+					int fd);
 
 #endif
