+++ libft/srcs/math/ft_factorial.c
@@ -6,13 +6,11 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/20 01:16:13 by abidolet          #+#    #+#             */
-/*   Updated: 2025/05/10 23:17:24 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/04 23:29:52 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
-
-size_t	ft_factorial(size_t n)
+unsigned long	ft_factorial(unsigned long n)
 {
 	if (n < 2)
 		return (1);
