+++ libft/srcs/mem/ft_memcmp.c
@@ -10,9 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
-
-int	ft_memcmp(const void *s1, const void *s2, size_t n)
+int	ft_memcmp(const void *s1, const void *s2, unsigned long n)
 {
 	while (n-- > 0)
 		if (*((unsigned char *)s1++) != *((unsigned char *)s2++))
