+++ libft/srcs/mem/ft_memcpy.c
@@ -6,13 +6,11 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:18:55 by abidolet          #+#    #+#             */
-/*   Updated: 2025/04/04 15:12:45 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/04 23:30:04 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
-
-void	*ft_memcpy(void *dest, const void *src, size_t n)
+void	*ft_memcpy(void *dest, const void *src, unsigned long n)
 {
 	if (!dest && !src)
 		return (dest);
