+++ libft/srcs/mem/ft_memset.c
@@ -6,13 +6,11 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:19:16 by abidolet          #+#    #+#             */
-/*   Updated: 2025/04/15 20:18:24 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/04 23:30:08 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
-
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, unsigned long n)
 {
 	while (n-- > 0)
 		*(unsigned char *)(s + n) = (unsigned char)c;
