+++ libft/srcs/str/ft_strlcpy.c
@@ -11,8 +11,9 @@
 /* ************************************************************************** */
 
 #include "libft/str.h"
+#include <stddef.h>
 
-size_t	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t	ft_strlcpy(char *dest, const char *src, unsigned long size)
 {
 	if (!dest || !src)
 		return (ft_strlen(src));
