+++ libft/srcs/str/ft_strncmp.c
@@ -10,9 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
-
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, unsigned long n)
 {
 	while (n-- && (*s1 || *s2) && *s1++ == *s2++)
 		;
